research: split variants into files, specify dynamic approval binding
variants.ts becomes variants/{approval,question,limit,challenge}.ts.
The approval variant now records the dynamic/ephemeral policy contract:
specs persist only durable facts (action, intentKey string,
responseAuthRequired); the response policy is late-bound per pass from
the live HarnessToolMap (bindApprovalPolicy, now implemented in seam.ts
— the authorizeCandidate lookup generalized). When the tool is absent,
the park-time responseAuthRequired flag decides: settle directly when
no policy was ever required, fail closed with policy-failed (row stays
open, answerable after redeploy) when one was.
